updateStatus
Conflicts:
	src/main/java/com/ctrip/zeus/restful/resource/StatusResource.java

+++ src/main/java/com/ctrip/zeus/service/status/impl/StatusServiceImpl.java
@@ -4,6 +4,8 @@ import com.ctrip.zeus.dal.core.*;
 import com.ctrip.zeus.service.status.StatusOffset;
 import com.ctrip.zeus.service.status.StatusService;
 import com.ctrip.zeus.status.entity.UpdateStatusItem;
+import com.netflix.config.DynamicBooleanProperty;
+import com.netflix.config.DynamicPropertyFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -23,7 +25,13 @@ public class StatusServiceImpl implements StatusService {
     private StatusServerDao statusServerDao;
     @Resource
     private StatusOffset statusOffset;
+    @Resource
+    private GroupServerStatusDao groupServerStatusDao;
+
     private final int OFFSET_LENGTH = 5;
+
+    private static DynamicBooleanProperty groupServerStatusNewEnable = DynamicPropertyFactory.getInstance().getBooleanProperty("group.server.status.new.version.enable", false);
+    private static DynamicBooleanProperty groupServerStatusOldDisable = DynamicPropertyFactory.getInstance().getBooleanProperty("group.server.status.old.version.disable", false);
 
 
     private Logger logger = LoggerFactory.getLogger(StatusServiceImpl.class);
@@ -37,45 +45,43 @@ public class StatusServiceImpl implements StatusService {
         }
         return allDownIps;
     }
+
     /*
     * group server is up while status is 0 .
     * */
-
     @Override
-    public Set<String> fetchGroupServersByVsIdsAndStatusOffset(Long[] vsIds, int offset, boolean status) throws Exception {
-        Set<String> result = new HashSet<>();
-        List<StatusGroupServerDo> statusGroupServerDos = statusGroupServerDao.findAllBySlbVirtualServerIds(vsIds, StatusGroupServerEntity.READSET_FULL);
-        int tmp = ~(1 << offset);
-        for (StatusGroupServerDo statusGroupServerDo : statusGroupServerDos) {
-            int tmpStatus = statusGroupServerDo.getStatus();
-            /*
-            * tmp == (tmpStatus|tmp) : 1. true  offsetStatus = true(0)  2. false: offsetStatus = false(1)
-            * offsetStatus equals (tmp == (tmpStatus|tmp))
-            * */
-            boolean offsetStatus = (tmp == (tmpStatus | tmp));
-            if (status == offsetStatus) {
-                result.add(statusGroupServerDo.getSlbVirtualServerId() + "_" + statusGroupServerDo.getGroupId() + "_" + statusGroupServerDo.getIp());
-            }
-        }
-        return result;
-    }
-
-    @Override
-    public Map<String, List<Boolean>> fetchGroupServersByVsIds(Long[] vsIds) throws Exception {
+    public Map<String, List<Boolean>> fetchGroupServerStatus(Long[] vsIds, Long[] groupIds) throws Exception {
         Map<String, List<Boolean>> result = new HashMap<>();
-        List<StatusGroupServerDo> statusGroupServerDos = statusGroupServerDao.findAllBySlbVirtualServerIds(vsIds, StatusGroupServerEntity.READSET_FULL);
-        int tmp = 1;
-        for (StatusGroupServerDo statusGroupServerDo : statusGroupServerDos) {
-            int tmpStatus = statusGroupServerDo.getStatus();
+        if (groupServerStatusNewEnable.get()) {
+            List<GroupServerStatusDo> groupServerStatusDos = groupServerStatusDao.findAllByGroupIds(groupIds, GroupServerStatusEntity.READSET_FULL);
+            int tmp = 1;
+            for (GroupServerStatusDo s : groupServerStatusDos) {
+                int tmpStatus = s.getStatus();
+                /*
+                * offset == 0 is true ; offset == 1 is false.
+                * */
+                List<Boolean> offset = new ArrayList<>(OFFSET_LENGTH);
+                for (int i = 0; i < OFFSET_LENGTH; i++) {
+                    offset.add(offset.size(), 0 == (tmpStatus & tmp));
+                    tmpStatus = tmpStatus >> 1;
+                }
+                result.put(s.getGroupId() + "_" + s.getIp(), offset);
+            }
+        } else {
+            List<StatusGroupServerDo> statusGroupServerDos = statusGroupServerDao.findAllBySlbVirtualServerIds(vsIds, StatusGroupServerEntity.READSET_FULL);
+            int tmp = 1;
+            for (StatusGroupServerDo statusGroupServerDo : statusGroupServerDos) {
+                int tmpStatus = statusGroupServerDo.getStatus();
             /*
             * offset == 0 is true ; offset == 1 is false.
             * */
-            List<Boolean> offset = new ArrayList<>(OFFSET_LENGTH);
-            for (int i = 0; i < OFFSET_LENGTH; i++) {
-                offset.add(offset.size(), 0 == (tmpStatus & tmp));
-                tmpStatus = tmpStatus >> 1;
+                List<Boolean> offset = new ArrayList<>(OFFSET_LENGTH);
+                for (int i = 0; i < OFFSET_LENGTH; i++) {
+                    offset.add(offset.size(), 0 == (tmpStatus & tmp));
+                    tmpStatus = tmpStatus >> 1;
+                }
+                result.put(statusGroupServerDo.getGroupId() + "_" + statusGroupServerDo.getIp(), offset);
             }
-            result.put(statusGroupServerDo.getSlbVirtualServerId() + "_" + statusGroupServerDo.getGroupId() + "_" + statusGroupServerDo.getIp(), offset);
         }
         return result;
     }
@@ -102,45 +108,86 @@ public class StatusServiceImpl implements StatusService {
         if (offset > OFFSET_LENGTH || offset < 0) {
             throw new Exception("offset of status should be [0-" + OFFSET_LENGTH + "]");
         }
-        for (String ip : ips) {
-            if (ip == null || ip.isEmpty()) {
-                continue;
+        if (!groupServerStatusOldDisable.get()) {
+            for (String ip : ips) {
+                if (ip == null || ip.isEmpty()) {
+                    continue;
+                }
+                StatusGroupServerDo data = new StatusGroupServerDo();
+                data.setSlbVirtualServerId(vsId)
+                        .setGroupId(groupId)
+                        .setIp(ip)
+                        .setCreatedTime(new Date());
+                int reset = ~(1 << offset);
+                int updatestatus = (status ? 0 : 1) << offset;
+                data.setReset(reset).setStatus(updatestatus);
+                statusGroupServerDao.updateStatus(data);
             }
-            StatusGroupServerDo data = new StatusGroupServerDo();
-            data.setSlbVirtualServerId(vsId)
-                    .setGroupId(groupId)
-                    .setIp(ip)
-                    .setCreatedTime(new Date());
-            int reset = ~(1 << offset);
-            int updatestatus = (status ? 0 : 1) << offset;
-            data.setReset(reset).setStatus(updatestatus);
-            statusGroupServerDao.updateStatus(data);
+        }
+        if (groupServerStatusNewEnable.get()) {
+            for (String ip : ips) {
+                if (ip == null || ip.isEmpty()) {
+                    continue;
+                }
+                GroupServerStatusDo data = new GroupServerStatusDo();
+                data.setGroupId(groupId);
+                data.setIp(ip);
+                data.setCreatedTime(new Date());
+                int reset = ~(1 << offset);
+                int updatestatus = (status ? 0 : 1) << offset;
+                data.setReset(reset).setStatus(updatestatus);
+                groupServerStatusDao.updateStatus(data);
+            }
         }
     }
 
     @Override
     public void updateStatus(List<UpdateStatusItem> items) throws Exception {
-        List<StatusGroupServerDo> updateDatas = new ArrayList<>();
-        for (UpdateStatusItem item : items) {
-            if (item.getOffset() > OFFSET_LENGTH || item.getOffset() < 0) {
-                throw new Exception("offset of status should be [0-" + OFFSET_LENGTH + "]");
-            }
-            for (String ip : item.getIpses()) {
-                if (ip == null || ip.isEmpty()) {
-                    continue;
+        if (!groupServerStatusOldDisable.get()) {
+            List<StatusGroupServerDo> updateDatas = new ArrayList<>();
+            for (UpdateStatusItem item : items) {
+                if (item.getOffset() > OFFSET_LENGTH || item.getOffset() < 0) {
+                    throw new Exception("offset of status should be [0-" + OFFSET_LENGTH + "]");
                 }
-                StatusGroupServerDo data = new StatusGroupServerDo();
-                data.setSlbVirtualServerId(item.getVsId())
-                        .setGroupId(item.getGroupId())
-                        .setIp(ip)
-                        .setCreatedTime(new Date());
-                int reset = ~(1 << item.getOffset());
-                int updatestatus = (item.isUp() ? 0 : 1) << item.getOffset();
-                data.setReset(reset).setStatus(updatestatus);
-                updateDatas.add(data);
+                for (String ip : item.getIpses()) {
+                    if (ip == null || ip.isEmpty()) {
+                        continue;
+                    }
+                    StatusGroupServerDo data = new StatusGroupServerDo();
+                    data.setSlbVirtualServerId(item.getVsId())
+                            .setGroupId(item.getGroupId())
+                            .setIp(ip)
+                            .setCreatedTime(new Date());
+                    int reset = ~(1 << item.getOffset());
+                    int updatestatus = (item.isUp() ? 0 : 1) << item.getOffset();
+                    data.setReset(reset).setStatus(updatestatus);
+                    updateDatas.add(data);
+                }
             }
+            statusGroupServerDao.batchUpdateStatus(updateDatas.toArray(new StatusGroupServerDo[]{}));
         }
-        statusGroupServerDao.batchUpdateStatus(updateDatas.toArray(new StatusGroupServerDo[]{}));
+        if (groupServerStatusNewEnable.get()) {
+            List<GroupServerStatusDo> updateDatas = new ArrayList<>();
+            for (UpdateStatusItem item : items) {
+                if (item.getOffset() > OFFSET_LENGTH || item.getOffset() < 0) {
+                    throw new Exception("offset of status should be [0-" + OFFSET_LENGTH + "]");
+                }
+                for (String ip : item.getIpses()) {
+                    if (ip == null || ip.isEmpty()) {
+                        continue;
+                    }
+                    GroupServerStatusDo data = new GroupServerStatusDo();
+                    data.setGroupId(item.getGroupId());
+                    data.setIp(ip);
+                    data.setCreatedTime(new Date());
+                    int reset = ~(1 << item.getOffset());
+                    int updatestatus = (item.isUp() ? 0 : 1) << item.getOffset();
+                    data.setReset(reset).setStatus(updatestatus);
+                    updateDatas.add(data);
+                }
+            }
+            groupServerStatusDao.batchUpdateStatus(updateDatas.toArray(new GroupServerStatusDo[]{}));
+        }
     }
 
 
@@ -154,28 +201,26 @@ public class StatusServiceImpl implements StatusService {
     }
 
     @Override
-    public Set<Long> findGroupIdByIp(String ip) throws Exception {
-        Set<Long> result = new HashSet<>();
-        List<StatusGroupServerDo> list = statusGroupServerDao.findAllByIp(ip, StatusGroupServerEntity.READSET_FULL);
-        if (list == null) {
-            return result;
-        }
-        for (StatusGroupServerDo statusGroupServerDo : list) {
-            result.add(statusGroupServerDo.getGroupId());
-        }
-        return result;
-    }
-
-    @Override
     public void groupServerStatusInit(Long groupId, Long[] vsIds, String[] ips) throws Exception {
         /*
         * only called in group resource api
         * */
-        for (Long vsId : vsIds) {
+        if (!groupServerStatusOldDisable.get()) {
+            for (Long vsId : vsIds) {
+                for (String ip : ips) {
+                    statusGroupServerDao.insert(new StatusGroupServerDo()
+                            .setGroupId(groupId)
+                            .setSlbVirtualServerId(vsId)
+                            .setIp(ip)
+                            .setStatus(statusOffset.getDefaultStatus())
+                            .setCreatedTime(new Date()));
+                }
+            }
+        }
+        if (groupServerStatusNewEnable.get()) {
             for (String ip : ips) {
-                statusGroupServerDao.insert(new StatusGroupServerDo()
+                groupServerStatusDao.insert(new GroupServerStatusDo()
                         .setGroupId(groupId)
-                        .setSlbVirtualServerId(vsId)
                         .setIp(ip)
                         .setStatus(statusOffset.getDefaultStatus())
                         .setCreatedTime(new Date()));
@@ -185,6 +230,11 @@ public class StatusServiceImpl implements StatusService {
 
     @Override
     public void cleanGroupServerStatus(Long groupId) throws Exception {
-        statusGroupServerDao.deleteByGroupId(new StatusGroupServerDo().setGroupId(groupId));
+        if (!groupServerStatusOldDisable.get()) {
+            statusGroupServerDao.deleteByGroupId(new StatusGroupServerDo().setGroupId(groupId));
+        }
+        if (groupServerStatusNewEnable.get()) {
+            groupServerStatusDao.deleteByGroupId(new GroupServerStatusDo().setGroupId(groupId));
+        }
     }
 }
